.travis: restore skip-setup
Forgot why I put this there in the first place in the last commit. I put
it there because we can't control Travis' variables so specifically as
we can the bots, like outdated xcode and xquartz versions.

Skipping the setup means we can actually get builds to pass, instead of
consta-fail.